+++ packages/web-visualization/src/chart/line/__stories__/LineChart.stories.tsx
@@ -1,10 +1,11 @@
 import { forwardRef, memo, useCallback, useEffect, useId, useMemo, useRef, useState } from 'react';
-import { assets } from '@coinbase/cds-common/internal/data/assets';
+import { assets, ethBackground } from '@coinbase/cds-common/internal/data/assets';
 import { candles as btcCandles } from '@coinbase/cds-common/internal/data/candles';
 import { prices } from '@coinbase/cds-common/internal/data/prices';
 import { sparklineInteractiveData } from '@coinbase/cds-common/internal/visualizations/SparklineInteractiveData';
 import { useTabsContext } from '@coinbase/cds-common/tabs/TabsContext';
 import type { TabValue } from '@coinbase/cds-common/tabs/useTabs';
+import { DataCard } from '@coinbase/cds-web/alpha/data-card/DataCard';
 import { ListCell } from '@coinbase/cds-web/cells';
 import { useBreakpoints } from '@coinbase/cds-web/hooks/useBreakpoints';
 import { Box, HStack, VStack } from '@coinbase/cds-web/layout';
@@ -23,7 +24,6 @@ import { m } from 'framer-motion';
 import {
   type AxisBounds,
   DefaultScrubberBeacon,
-  DefaultScrubberLabel,
   defaultTransition,
   PeriodSelector,
   PeriodSelectorActiveIndicator,
@@ -31,7 +31,6 @@ import {
   projectPoint,
   Scrubber,
   type ScrubberBeaconProps,
-  type ScrubberLabelProps,
   type ScrubberRef,
   useCartesianChartContext,
   useScrubberContext,
@@ -701,41 +700,6 @@ function StylingScrubber() {
       }}
     >
       <Scrubber idlePulse LineComponent={SolidLine} seriesIds={['pageViews']} />
-    </LineChart>
-  );
-}
-
-function HideBeaconLabels() {
-  const pageViews = [2400, 1398, 9800, 3908, 4800, 3800, 4300];
-  const uniqueVisitors = [4000, 3000, 2000, 2780, 1890, 2390, 3490];
-
-  return (
-    <LineChart
-      enableScrubbing
-      legend
-      showArea
-      height={{ base: 200, tablet: 225, desktop: 250 }}
-      inset={{ top: 60 }}
-      series={[
-        {
-          id: 'pageViews',
-          data: pageViews,
-          color: 'var(--color-accentBoldGreen)',
-          label: 'Page Views',
-        },
-        {
-          id: 'uniqueVisitors',
-          data: uniqueVisitors,
-          color: 'var(--color-accentBoldPurple)',
-          label: 'Unique Visitors',
-        },
-      ]}
-    >
-      <Scrubber
-        hideBeaconLabels
-        labelElevated
-        label={(dataIndex: number) => `Day ${dataIndex + 1}`}
-      />
     </LineChart>
   );
 }
@@ -1585,110 +1549,6 @@ function MonotoneAssetPrice() {
   );
 }
 
-function CustomLabelComponent() {
-  const CustomLabelComponent = memo((props: ScrubberLabelProps) => {
-    const { drawingArea } = useCartesianChartContext();
-
-    if (!drawingArea) return;
-
-    return (
-      <DefaultScrubberLabel
-        {...props}
-        elevated
-        background="var(--color-bgPrimary)"
-        color="var(--color-bgPrimaryWash)"
-        dy={32}
-        fontWeight="label1"
-        y={drawingArea.y + drawingArea.height}
-      />
-    );
-  });
-  return (
-    <LineChart
-      enableScrubbing
-      showArea
-      height={{ base: 150, tablet: 200, desktop: 250 }}
-      inset={{ top: 16, bottom: 64 }}
-      series={[
-        {
-          id: 'prices',
-          data: sampleData,
-        },
-      ]}
-    >
-      <Scrubber
-        LabelComponent={CustomLabelComponent}
-        label={(dataIndex: number) => `Day ${dataIndex + 1}`}
-      />
-    </LineChart>
-  );
-}
-
-function CustomBeaconStroke() {
-  const backgroundColor = 'rgb(var(--red40))';
-  const foregroundColor = 'rgb(var(--gray0))';
-
-  return (
-    <Box borderRadius={300} padding={2} style={{ background: backgroundColor }}>
-      <LineChart
-        enableScrubbing
-        showArea
-        height={{ base: 150, tablet: 200, desktop: 250 }}
-        series={[
-          {
-            id: 'prices',
-            data: sampleData,
-            color: foregroundColor,
-          },
-        ]}
-      >
-        <Scrubber
-          hideOverlay
-          idlePulse
-          beaconStroke={backgroundColor}
-          lineStroke={foregroundColor}
-        />
-      </LineChart>
-    </Box>
-  );
-}
-
-function CustomBeaconSize() {
-  const InvertedBeacon = useMemo(
-    () => (props: ScrubberBeaconProps) => (
-      <DefaultScrubberBeacon
-        {...props}
-        color="var(--color-bg)"
-        radius={5}
-        stroke="var(--color-fg)"
-        strokeWidth={3}
-      />
-    ),
-    [],
-  );
-
-  return (
-    <LineChart
-      enableScrubbing
-      showArea
-      showYAxis
-      height={{ base: 150, tablet: 200, desktop: 250 }}
-      series={[
-        {
-          id: 'prices',
-          data: sampleData,
-          color: 'var(--color-fg)',
-        },
-      ]}
-      yAxis={{
-        showGrid: true,
-      }}
-    >
-      <Scrubber BeaconComponent={InvertedBeacon} />
-    </LineChart>
-  );
-}
-
 export const All = () => {
   return (
     <VStack gap={2}>
@@ -1884,17 +1744,8 @@ export const All = () => {
       <Example title="Forecast Asset Price">
         <ForecastAssetPrice />
       </Example>
-      <Example title="Custom Label Component">
-        <CustomLabelComponent />
-      </Example>
-      <Example title="Hide Beacon Labels">
-        <HideBeaconLabels />
-      </Example>
-      <Example title="Custom Beacon Stroke">
-        <CustomBeaconStroke />
-      </Example>
-      <Example title="Custom Beacon Size">
-        <CustomBeaconSize />
+      <Example title="In DataCard">
+        <DataCardWithLineChart />
       </Example>
     </VStack>
   );
@@ -2033,3 +1884,120 @@ export const Transitions = () => {
 
   return <CustomTransitionsChart />;
 };
+function DataCardWithLineChart() {
+  const exampleThumbnail = (
+    <RemoteImage
+      accessibilityLabel="Ethereum"
+      shape="circle"
+      size="l"
+      source={ethBackground}
+      testID="thumbnail"
+    />
+  );
+
+  const getLineChartSeries = () => [
+    {
+      id: 'price',
+      data: prices.slice(0, 30).map((price: string) => parseFloat(price)),
+      color: 'var(--color-accentBoldBlue)',
+    },
+  ];
+
+  const lineChartSeries = useMemo(() => getLineChartSeries(), []);
+  const lineChartSeries2 = useMemo(() => getLineChartSeries(), []);
+  const ref = useRef<HTMLAnchorElement>(null);
+
+  return (
+    <VStack gap={2}>
+      <DataCard
+        layout="vertical"
+        subtitle="Price trend"
+        thumbnail={exampleThumbnail}
+        title="Line Chart Card"
+      >
+        <LineChart
+          showArea
+          accessibilityLabel="Ethereum price chart"
+          areaType="dotted"
+          height={120}
+          inset={0}
+          series={lineChartSeries}
+        />
+      </DataCard>
+      <DataCard
+        layout="vertical"
+        subtitle="Price trend"
+        thumbnail={exampleThumbnail}
+        title="Line Chart with Tag"
+        titleAccessory={
+          <Text dangerouslySetColor="rgb(var(--green70))" font="label1">
+            ↗ 25.25%
+          </Text>
+        }
+      >
+        <LineChart
+          showArea
+          accessibilityLabel="Ethereum price chart"
+          areaType="dotted"
+          height={100}
+          inset={0}
+          series={lineChartSeries}
+        />
+      </DataCard>
+      <DataCard
+        ref={ref}
+        renderAsPressable
+        as="a"
+        href="https://www.coinbase.com"
+        layout="vertical"
+        subtitle="Clickable line chart card"
+        target="_blank"
+        thumbnail={exampleThumbnail}
+        title="Actionable Line Chart"
+        titleAccessory={
+          <Text dangerouslySetColor="rgb(var(--green70))" font="label1">
+            ↗ 25.25%
+          </Text>
+        }
+      >
+        <LineChart
+          showArea
+          accessibilityLabel="Ethereum price chart"
+          areaType="dotted"
+          height={120}
+          inset={0}
+          series={lineChartSeries}
+        />
+      </DataCard>
+
+      <DataCard
+        layout="vertical"
+        subtitle="Price trend"
+        thumbnail={
+          <RemoteImage
+            accessibilityLabel="Bitcoin"
+            shape="circle"
+            size="l"
+            source={assets.btc.imageUrl}
+            testID="thumbnail"
+          />
+        }
+        title="Card with Line Chart"
+        titleAccessory={
+          <Text dangerouslySetColor="rgb(var(--green70))" font="label1">
+            ↗ 25.25%
+          </Text>
+        }
+      >
+        <LineChart
+          showArea
+          accessibilityLabel="Price chart"
+          areaType="dotted"
+          height={100}
+          inset={0}
+          series={lineChartSeries2}
+        />
+      </DataCard>
+    </VStack>
+  );
+}